requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0
python-dotenv==1.0.0
python-dateutil==2.8.2
langchain-community>=0.0.10
requests>=2.31.0
